Add LTL formula for 'request for floor serviced eventually'
This commit updates the elevator-system.pml file to add and
check the property as mentioned. A new state has been added
to the elevator controller to halt the elevator while change
of direction of motion. A corresponding change has been made
in the elevator process to handshake with this direction
changing halt, in the wait_for_move state. This commit passes
the check for the mentioned LTL formula.